+++ agora-specs/Spec/Stake.hs
@@ -15,6 +15,8 @@ import Agora.Stake (
   StakeRedeemer (DepositWithdraw),
  )
 import Agora.Stake.Scripts (stakePolicy, stakeValidator)
+import Data.Bool (Bool (..))
+import Data.Maybe (Maybe (..))
 import Sample.Stake (
   DepositWithdrawExample (
     DepositWithdrawExample,
@@ -30,6 +32,7 @@ import Sample.Stake qualified as Stake (
   stakeCreationWrongDatum,
   stakeDepositWithdraw,
  )
+import Sample.Stake.SetDelegate qualified as SetDelegate
 import Test.Specification (
   SpecificationTree,
   group,
@@ -67,20 +70,47 @@ specs =
       [ validatorSucceedsWith
           "stakeDepositWithdraw deposit"
           (stakeValidator Stake.stake)
-          (toDatum $ StakeDatum 100_000 signer [])
+          (toDatum $ StakeDatum 100_000 signer Nothing [])
           (toDatum $ DepositWithdraw 100_000)
           (Stake.stakeDepositWithdraw $ DepositWithdrawExample {startAmount = 100_000, delta = 100_000})
       , validatorSucceedsWith
           "stakeDepositWithdraw withdraw"
           (stakeValidator Stake.stake)
-          (toDatum $ StakeDatum 100_000 signer [])
+          (toDatum $ StakeDatum 100_000 signer Nothing [])
           (toDatum $ DepositWithdraw $ negate 100_000)
           (Stake.stakeDepositWithdraw $ DepositWithdrawExample {startAmount = 100_000, delta = negate 100_000})
       , validatorFailsWith
           "stakeDepositWithdraw negative GT"
           (stakeValidator Stake.stake)
-          (toDatum $ StakeDatum 100_000 signer [])
+          (toDatum $ StakeDatum 100_000 signer Nothing [])
           (toDatum $ DepositWithdraw 1_000_000)
           (Stake.stakeDepositWithdraw $ DepositWithdrawExample {startAmount = 100_000, delta = negate 1_000_000})
+      , group
+          "set delegate"
+          [ SetDelegate.mkTestCase
+              "override existing delegate"
+              SetDelegate.overrideExistingDelegateParameters
+              True
+          , SetDelegate.mkTestCase
+              "remove existing delegate"
+              SetDelegate.clearDelegateParameters
+              True
+          , SetDelegate.mkTestCase
+              "set delegate to something"
+              SetDelegate.setDelegateParameters
+              True
+          , SetDelegate.mkTestCase
+              "owner doesn't sign the transaction"
+              SetDelegate.ownerDoesntSignParameters
+              False
+          , SetDelegate.mkTestCase
+              "delegate to the owner"
+              SetDelegate.delegateToOwnerParameters
+              False
+          , SetDelegate.mkTestCase
+              "invalid output stake"
+              SetDelegate.invalidOutputStakeDatumParameters
+              False
+          ]
       ]
   ]